Close #33 - Create UBO Abstraction
Created a simple abstraction around a Uniform Buffer Object. It is templated to hold any kind of data, although the data should all be stored locally within DataType for it to function properly. Data can be updated with the various update functions, and you can bind the buffer to a given UBO binding point with .bind(n)