Fix stupid bug in renderer where VAO was never initialized
So, I was smart enough to initialize the VAO in the default Ctor that is never called, nor used, and so my vertex data never bothered to show up anywhere. Thankfully I found it and can now work to revert some of the other changes I made in order to test this. At least rendering is fairly clean now, binding only the VAO as was initially planned.