Profile code and tweak functions for performance
Updated ImGui to not check for mouse cursor changes every frame as this was 13% CPU time on Windows, yielding ~700 FPS.

Stopped mapping and unmapping the instance buffer every frame when drawing, and instead mapped it persistently throughout the application's lifetime for an approx gain of ~600 FPS.

Went from 3200 to 4700 FPS on Windows. Profiling was done in Visual Studio 2017 x64.